hdl/top/ml_605/*/dbe_bpm_dsp.vhd: change delay FMC ADC delay type
The previous default value was "VARIABLE", which has the advantage
of enabling intiial delay values in the UCF. However, changing
the value dynamically (in software) is not much pratical as we
would need to update the tap value one by one, incrementally.

The current value "VAR_LOADBLE" does not accept the initial
delay value in the UCF, but changing them in software is easier.